fix imports, add test for medgen
